+++ magda-indexer/src/main/scala/au/csiro/data61/magda/indexer/Protocols.scala
@@ -1,0 +1,9 @@
+package au.csiro.data61.magda.indexer
+
+import au.csiro.data61.magda.indexer.search.SearchIndexer.IndexResult
+import spray.json.{DefaultJsonProtocol, RootJsonFormat}
+
+trait Protocols extends DefaultJsonProtocol {
+  implicit val IndexResultFormat: RootJsonFormat[IndexResult] =
+    jsonFormat5(IndexResult.apply)
+}